Move `MockInMemoryStorage` to be sharable.
