+++ WooCommerce/WooCommerceTests/Model/BetaFeaturesTests.swift
@@ -2,7 +2,7 @@ import XCTest
 import Storage
 @testable import WooCommerce
 
-class BetaFeaturesTests: XCTestCase {
+final class BetaFeaturesTests: XCTestCase {
     var appSettings: GeneralAppSettingsStorage!
 
     override func setUpWithError() throws {
@@ -30,28 +30,5 @@ class BetaFeaturesTests: XCTestCase {
         enabledBinding.wrappedValue = true
         let enabled = appSettings.betaFeatureEnabled(.viewAddOns)
         XCTAssertEqual(enabled, true)
-    }
-}
-
-private final class MockInMemoryStorage: FileStorage {
-    private(set) var data: [URL: Codable] = [:]
-
-    func data<T>(for fileURL: URL) throws -> T where T: Decodable {
-        guard let data = data[fileURL] as? T else {
-            throw Errors.readFailed
-        }
-        return data
-    }
-
-    func write<T>(_ data: T, to fileURL: URL) throws where T: Encodable {
-        self.data[fileURL] = data as? Codable
-    }
-
-    func deleteFile(at fileURL: URL) throws {
-        data.removeValue(forKey: fileURL)
-    }
-
-    enum Errors: Error {
-        case readFailed
     }
 }
